test: split config source and path tests
